Use global EventEmitter for listener management
Note: this is a breaking change requiring a major version bump.

From an API perspective, the biggest change is that the onExit callback
is not provided with the exit code corresponding to `127 + signal code`.
The reason for this is the same as for why it is not provided in Node's
ChildProcess close & exit events: the numbers vary across operating
systems, and are set in libc.  Thus, it's easier to just return `null`
as the exit code in those cases, and provide the signal only.

Instead of adding handlers directly to all of the signals, and using the
maxListeners option, with this change only a single emitter is used, so
that we only ever add 1 signal handler per signal.

In order to handle cases where the module might be loaded multiple
times, the emitter is hung on the process object at
`process.__signal_exit_emitter__`.  This object then has a `count`
property which keeps track of the number of times signal-exit has loaded
handlers.

The onSignalExit function now returns an `unwrap` method which will
remove the handler that the user provided.  When all handlers are
removed, it also then removed all of the signal event handlers and
process.exit handler.

Getting accurate coverage of this module required jumping through a few
hoops.  The unwrap tests in particular have to do some bizarre things to
pull nyc's handlers out of the way (and then re-add them later).  And,
any of the lines in the direct stack of emitting the `afterexit` event
will never be coverage-friendly, since they cannot complete until after
coverage has been finalized and written out.

With a few ignored lines and dirty tricks, however, coverage is 100%.

The biggest downside to this approach is that it does rely on a global
object hung on process.  At the very least, it's an extremely simple
shared interface, and EventEmitter is one of the most stable
abstractions in Node.