clobber reallyExit to handle exit() in exit event
If a user does `process.exit(n)` in a `process.on('exit')` handler, then
it was previously not detecting that the exit code had changed.

I don't see any way to handle this except by monkey-patching
process.reallyExit.  Since it only emits 'exit' once (and only with the
original exit code, not the final one), it's impossible to rely on the
argument to `process.emit('exit')` as a valid source of truth.

However, we *do* still need to listen to `process.on('exit')`, because
`process.reallyExit()` is not called in the event of a graceful
termination with a 0 status code.

Note: requires link-installing all the packages, so that the nyc in use
is the one using this version of this module.