More properly handle global.process mutating
The exit handler was set to a no-op whenever the process we grabbed at
the start didn't match global.process later on.

While this is safe, and fixed the initial failing, it is overly
conservative, and caused a test in npm/cli to break unnecessarily.  What
we actually care about is that the process we have is set up with all
the things we actually care about using.

This commit replaces the `process === global.process` checks with a
duck-typing function that ensures we won't throw unnecessarily, but
still allows these methods to be overridden in a new global process
object.